wakeup_backend/requirements.txt
Django===4.2
djangorestframework
djangorestframework-simplejwt[crypto]===5.3.1
django-cors-headers
psycopg2.binary
django-auth-ldap
python-ldap
scapy
wakeonlan
pythonping
gunicorn